+++ packages/lunora/CHANGELOG.md
@@ -1,3 +1,10 @@
+## lunorash [1.0.0-alpha.238](https://github.com/anolilab/lunora/compare/lunorash@1.0.0-alpha.237...lunorash@1.0.0-alpha.238) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/cli:** upgraded to 1.0.0-alpha.235
+
 ## lunorash [1.0.0-alpha.237](https://github.com/anolilab/lunora/compare/lunorash@1.0.0-alpha.236...lunorash@1.0.0-alpha.237) (2026-09-07)