+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/svelte [1.0.0-alpha.123](https://github.com/anolilab/lunora/compare/@lunora/svelte@1.0.0-alpha.122...@lunora/svelte@1.0.0-alpha.123) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.88
+
 ## @lunora/svelte [1.0.0-alpha.122](https://github.com/anolilab/lunora/compare/@lunora/svelte@1.0.0-alpha.121...@lunora/svelte@1.0.0-alpha.122) (2026-09-07)